Year 2 (2022-2023)/COMP0016 - Systems Engineering/Group Portfolio/code/src/MI_v3.2/requirements.txt
absl-py==1.3.0
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cmake==3.24.2
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
exceptiongroup==1.0.0
fonttools==4.38.0
idna==3.4
imutils==0.5.4
iniconfig==1.1.1
install==1.3.5
kiwisolver==1.4.4
matplotlib==3.6.1
mediapipe==0.8.9.1
MouseInfo==0.1.3
Nuitka==1.5.3
numpy==1.19.5
opencv-contrib-python==4.5.3.56
openvino==2021.4.2
orderedset==2.0.3
packaging==21.3
pandas==1.4.4
Pillow==9.3.0
pluggy==1.0.0
protobuf==3.20.0
PyAutoGUI==0.9.53
pycparser==2.21
PyDirectInput==1.0.4
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytest==7.2.0
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2022.6
pywin32==304
pyzmq==25.0.0
requests==2.28.1
scipy==1.9.3
six==1.16.0
sounddevice==0.4.5
srt==3.5.2
text2digits==0.1.0
tomli==2.0.1
tqdm==4.64.1
urllib3==1.26.12
vgamepad==0.0.8
vosk==0.3.44
websockets==10.4